refactor(web): migrate to core/IssuesHint to TypeScript
